refactor(models): split k8s backend into engine reconcilers
Extract the Kubernetes reconciliation logic out of the monolithic
K8sNimOperatorServiceBackend into a reconciler hierarchy under
reconcilers/:

* BaseReconciler (base.py): the reconciler interface plus the shared,
  engine-agnostic Kubernetes status helpers (pod log fetch, crash-loop
  detection, pod-status drill-down, pending-timeout/crash-loop error
  builders) and the idempotent 404-tolerant single-object delete.
* NimOperatorReconciler (nim_operator.py): emits NIMService / NIMCache
  CRs for the in-cluster operator to reconcile.
* K8sReconciler (k8s.py): emits native Kubernetes objects directly
  (PVC / weight-puller Job / Deployment / Service) for the vLLM engine
  and drives the staged rollout itself.

The ServiceBackend now owns only the nemo_platform SDK and API-object
work: it resolves all reconciler inputs into a ResolvedDeployment,
selects the reconciler by config.engine, and delegates. Delete asks both
reconcilers (no engine context) and aggregates; list unions both.

This is behavior-neutral: the NIM operator path is moved verbatim and
the vLLM path is unchanged. Tests are retargeted at the reconciler
modules/instances; the shared status-helper unit tests now exercise
BaseReconciler directly.
