+++ services/core/models/tests/unit/controllers/test_k8s_nim_operator_backend.py
@@ -18,9 +18,12 @@ from nmp.core.models.controllers.backends.backends import DeploymentStatusUpdate
 from nmp.core.models.controllers.backends.common import deployment_elapsed_seconds, format_duration
 from nmp.core.models.controllers.backends.k8s_nim_operator import K8sNimOperatorServiceBackend
 from nmp.core.models.controllers.backends.k8s_nim_operator.config import K8sNimOperatorConfig
+from nmp.core.models.controllers.backends.k8s_nim_operator.reconcilers.base import BaseReconciler
 from pydantic import ValidationError
 
 _K8S_BACKEND_MODULE = "nmp.core.models.controllers.backends.k8s_nim_operator.backend"
+_RECON_BASE_MODULE = "nmp.core.models.controllers.backends.k8s_nim_operator.reconcilers.base"
+_RECON_K8S_MODULE = "nmp.core.models.controllers.backends.k8s_nim_operator.reconcilers.k8s"
 
 
 # ---------------------------------------------------------------------------
@@ -93,8 +96,8 @@ def _mock_pod_backend(k8s_backend, pod=None, *, pod_logs=""):
     mock_core_v1.read_namespaced_pod_log.return_value = pod_logs
 
     with (
-        patch(f"{_K8S_BACKEND_MODULE}.k8s_client.AppsV1Api", return_value=mock_apps_v1),
-        patch(f"{_K8S_BACKEND_MODULE}.k8s_client.CoreV1Api", return_value=mock_core_v1),
+        patch(f"{_RECON_BASE_MODULE}.k8s_client.AppsV1Api", return_value=mock_apps_v1),
+        patch(f"{_RECON_BASE_MODULE}.k8s_client.CoreV1Api", return_value=mock_core_v1),
     ):
         yield mock_apps_v1, mock_core_v1
 
@@ -183,6 +186,77 @@ def k8s_backend(mock_nmp_sdk, mock_k8s_config):
     )
 
 
+def _sync_reconcilers(backend):
+    """Propagate the backend's (test-mocked) k8s state onto its reconcilers.
+
+    Tests assign mock clients/config/namespace onto the backend *after*
+    construction (``backend._dynamic_client = MagicMock()`` etc.). Reconciliation
+    logic now lives on the two reconcilers, which captured their own clients at
+    ``init()`` time. This helper re-points the reconcilers at whatever the test
+    set on the backend so delegation exercises the test's mocks. Call it after
+    setting up the backend's ``_dynamic_client`` / ``_core_v1`` / ``_apps_v1`` /
+    ``_batch_v1`` / ``_backend_config`` / ``_k8s_namespace`` / ``_k8s_client``.
+    """
+    nim = backend._nim_reconciler
+    k8s = backend._k8s_reconciler
+    namespace = backend._k8s_namespace
+    config = backend._backend_config
+    client = backend._k8s_client
+
+    if nim is not None:
+        nim._k8s_namespace = namespace
+        nim._backend_config = config
+        nim._k8s_client = client
+        if backend._dynamic_client is not None:
+            nim._dynamic_client = backend._dynamic_client
+    if k8s is not None:
+        k8s._k8s_namespace = namespace
+        k8s._backend_config = config
+        k8s._k8s_client = client
+        if getattr(backend, "_core_v1", None) is not None:
+            k8s._core_v1 = backend._core_v1
+        if getattr(backend, "_apps_v1", None) is not None:
+            k8s._apps_v1 = backend._apps_v1
+        if getattr(backend, "_batch_v1", None) is not None:
+            k8s._batch_v1 = backend._batch_v1
+    return backend
+
+
+class _StatusHelperReconciler(BaseReconciler):
+    """Minimal concrete BaseReconciler for unit-testing the shared status helpers.
+
+    The shared pod-status / log / crash-loop helpers live on ``BaseReconciler``;
+    these tests exercise them directly without needing a full engine reconciler.
+    """
+
+    async def create(self, resolved):  # pragma: no cover - not exercised
+        raise NotImplementedError
+
+    async def update(self, resolved):  # pragma: no cover - not exercised
+        raise NotImplementedError
+
+    async def get_status(self, resolved):  # pragma: no cover - not exercised
+        raise NotImplementedError
+
+    async def get_status_orphan(self, deployment, resource_name):  # pragma: no cover
+        raise NotImplementedError
+
+    async def delete(self, workspace, name):  # pragma: no cover - not exercised
+        raise NotImplementedError
+
+    async def list_managed_deployment_names(self):  # pragma: no cover - not exercised
+        raise NotImplementedError
+
+
+def _status_helper_reconciler(*, namespace="default", backend_config=None, k8s_client_=None):
+    """Build a BaseReconciler exposing the shared status helpers for direct tests."""
+    return _StatusHelperReconciler(
+        k8s_client_=k8s_client_ if k8s_client_ is not None else MagicMock(),
+        backend_config=backend_config if backend_config is not None else K8sNimOperatorConfig(),
+        k8s_namespace=namespace,
+    )
+
+
 @pytest.fixture
 def sample_deployment():
     """Create a sample ModelDeployment for testing.
@@ -223,11 +297,14 @@ async def test_k8s_backend_create_model_deployment(k8s_backend, sample_deploymen
     k8s_backend._dynamic_client.resources.get.return_value = mock_resource
 
     # Mock the compile_nimservice function to avoid validation issues
-    with patch("nmp.core.models.controllers.backends.k8s_nim_operator.backend.compile_nimservice") as mock_compile:
+    with patch(
+        "nmp.core.models.controllers.backends.k8s_nim_operator.reconcilers.nim_operator.compile_nimservice"
+    ) as mock_compile:
         mock_nimservice = MagicMock()
         mock_nimservice.model_dump.return_value = {"apiVersion": "apps.nvidia.com/v1alpha1", "kind": "NIMService"}
         mock_compile.return_value = mock_nimservice
 
+        _sync_reconcilers(k8s_backend)
         status_update = await k8s_backend.create_model_deployment(sample_deployment, sample_config)
 
         # Verify compile_nimservice was called
@@ -256,11 +333,14 @@ async def test_k8s_backend_update_model_deployment(k8s_backend, sample_deploymen
     k8s_backend._dynamic_client.resources.get.return_value = mock_resource
 
     # Mock the compile_nimservice function to avoid validation issues
-    with patch("nmp.core.models.controllers.backends.k8s_nim_operator.backend.compile_nimservice") as mock_compile:
+    with patch(
+        "nmp.core.models.controllers.backends.k8s_nim_operator.reconcilers.nim_operator.compile_nimservice"
+    ) as mock_compile:
         mock_nimservice = MagicMock()
         mock_nimservice.model_dump.return_value = {"apiVersion": "apps.nvidia.com/v1alpha1", "kind": "NIMService"}
         mock_compile.return_value = mock_nimservice
 
+        _sync_reconcilers(k8s_backend)
         status_update = await k8s_backend.update_model_deployment(sample_deployment, sample_config)
 
         # Verify compile_nimservice was called
@@ -282,6 +362,7 @@ async def test_k8s_backend_get_model_deployment_status(k8s_backend, sample_deplo
 
     k8s_backend._dynamic_client.resources.get.return_value = _make_nimservice_mock("Ready")
 
+    _sync_reconcilers(k8s_backend)
     status_update = await k8s_backend.get_model_deployment_status(sample_deployment)
 
     assert status_update is not None
@@ -301,6 +382,7 @@ async def test_k8s_backend_get_status_nimservice_not_found(k8s_backend, sample_d
     mock_resource.get.side_effect = k8s_dynamic_exceptions.NotFoundError(MagicMock())
     k8s_backend._dynamic_client.resources.get.return_value = mock_resource
 
+    _sync_reconcilers(k8s_backend)
     status_update = await k8s_backend.get_model_deployment_status(sample_deployment)
 
     assert status_update is not None
@@ -319,12 +401,13 @@ async def test_k8s_backend_get_status_nimservice_not_ready(k8s_backend, sample_d
     k8s_backend._dynamic_client.resources.get.return_value = _make_nimservice_mock("NotReady")
 
     with patch.object(
-        k8s_backend,
+        k8s_backend._nim_reconciler,
         "_get_pod_status_from_deployment",
         return_value=DeploymentStatusUpdate(
             status="PENDING", status_message="Waiting for NIMService to become ready", host_url=None
         ),
     ):
+        _sync_reconcilers(k8s_backend)
         status_update = await k8s_backend.get_model_deployment_status(sample_deployment)
 
         assert status_update is not None
@@ -347,6 +430,7 @@ async def test_k8s_backend_get_status_nimservice_crash_loop_backoff(k8s_backend,
     pod = _make_pod(restart_count=5, waiting_reason="CrashLoopBackOff")
 
     with _mock_pod_backend(k8s_backend, pod=pod, pod_logs="ERROR: model failed to load"):
+        _sync_reconcilers(k8s_backend)
         status_update = await k8s_backend.get_model_deployment_status(sample_deployment)
 
         assert status_update is not None
@@ -372,6 +456,7 @@ async def test_k8s_backend_get_status_nimservice_pod_restarts_below_threshold(k8
     pod = _make_pod(restart_count=2)
 
     with _mock_pod_backend(k8s_backend, pod=pod):
+        _sync_reconcilers(k8s_backend)
         status_update = await k8s_backend.get_model_deployment_status(sample_deployment)
 
         assert status_update is not None
@@ -393,6 +478,7 @@ async def test_k8s_backend_get_status_nimservice_pod_running_after_restarts(k8s_
     pod = _make_pod(restart_count=5)  # No waiting_reason → running
 
     with _mock_pod_backend(k8s_backend, pod=pod):
+        _sync_reconcilers(k8s_backend)
         status_update = await k8s_backend.get_model_deployment_status(sample_deployment)
 
         assert status_update is not None
@@ -413,6 +499,7 @@ async def test_k8s_backend_delete_model_deployment(k8s_backend, sample_deploymen
     mock_resource = MagicMock()
     k8s_backend._dynamic_client.resources.get.return_value = mock_resource
 
+    _sync_reconcilers(k8s_backend)
     status_update = await k8s_backend.delete_model_deployment(sample_deployment.workspace, sample_deployment.name)
 
     # Verify status update returned
@@ -434,6 +521,7 @@ async def test_k8s_backend_delete_model_deployment_with_secret(k8s_backend, samp
     mock_resource = MagicMock()
     k8s_backend._dynamic_client.resources.get.return_value = mock_resource
 
+    _sync_reconcilers(k8s_backend)
     status_update = await k8s_backend.delete_model_deployment(sample_deployment.workspace, sample_deployment.name)
 
     # Verify status update returned
@@ -454,6 +542,7 @@ async def test_k8s_backend_delete_model_deployment_without_secret(k8s_backend, s
     mock_resource = MagicMock()
     k8s_backend._dynamic_client.resources.get.return_value = mock_resource
 
+    _sync_reconcilers(k8s_backend)
     status_update = await k8s_backend.delete_model_deployment(sample_deployment.workspace, sample_deployment.name)
 
     # Verify status update returned
@@ -478,6 +567,7 @@ async def test_delete_attempts_all_resource_types_and_tolerates_404(k8s_backend,
     k8s_backend._batch_v1.delete_namespaced_job.side_effect = notfound
     k8s_backend._core_v1.delete_namespaced_persistent_volume_claim.side_effect = notfound
 
+    _sync_reconcilers(k8s_backend)
     result = await k8s_backend.delete_model_deployment("default", "qwen")
 
     assert result.status == "DELETED"
@@ -505,6 +595,7 @@ async def test_delete_real_failure_surfaces_error_but_attempts_all(k8s_backend, 
     k8s_backend._batch_v1.delete_namespaced_job.side_effect = notfound
     k8s_backend._core_v1.delete_namespaced_persistent_volume_claim.side_effect = notfound
 
+    _sync_reconcilers(k8s_backend)
     result = await k8s_backend.delete_model_deployment("default", "qwen")
 
     assert result.status == "ERROR"
@@ -524,6 +615,7 @@ async def test_delete_forbidden_cr_does_not_block_vllm_cleanup(k8s_backend, samp
     cr_api.delete.side_effect = k8s_dynamic_exceptions.ForbiddenError(MagicMock(status=403))
     k8s_backend._dynamic_client.resources.get.return_value = cr_api
 
+    _sync_reconcilers(k8s_backend)
     result = await k8s_backend.delete_model_deployment("default", "qwen")
 
     assert result.status == "ERROR"
@@ -582,11 +674,14 @@ async def test_k8s_backend_create_when_nimservice_already_exists(k8s_backend, sa
     k8s_backend._dynamic_client.resources.get.return_value = mock_resource
 
     # Mock the compile_nimservice function
-    with patch("nmp.core.models.controllers.backends.k8s_nim_operator.backend.compile_nimservice") as mock_compile:
+    with patch(
+        "nmp.core.models.controllers.backends.k8s_nim_operator.reconcilers.nim_operator.compile_nimservice"
+    ) as mock_compile:
         mock_nimservice = MagicMock()
         mock_nimservice.model_dump.return_value = {"apiVersion": "apps.nvidia.com/v1alpha1", "kind": "NIMService"}
         mock_compile.return_value = mock_nimservice
 
+        _sync_reconcilers(k8s_backend)
         status_update = await k8s_backend.create_model_deployment(sample_deployment, sample_config)
 
         # Verify status update returned PENDING (not ERROR)
@@ -672,11 +767,14 @@ async def test_create_model_deployment_with_sft_model(k8s_backend, sample_deploy
         "nmp.core.models.controllers.backends.k8s_nim_operator.nimservice_compiler.get_platform_config",
         return_value=platform_config,
     ):
-        with patch("nmp.core.models.controllers.backends.k8s_nim_operator.backend.compile_nimservice") as mock_compile:
+        with patch(
+            "nmp.core.models.controllers.backends.k8s_nim_operator.reconcilers.nim_operator.compile_nimservice"
+        ) as mock_compile:
             mock_nimservice = MagicMock()
             mock_nimservice.model_dump.return_value = {"apiVersion": "apps.nvidia.com/v1alpha1", "kind": "NIMService"}
             mock_compile.return_value = mock_nimservice
 
+            _sync_reconcilers(k8s_backend)
             status_update = await k8s_backend.create_model_deployment(
                 sample_deployment, sample_config, sft_model_entity
             )
@@ -768,11 +866,14 @@ async def test_create_model_deployment_with_files_service_model_triggers_nimcach
         "nmp.core.models.controllers.backends.k8s_nim_operator.nimservice_compiler.get_platform_config",
         return_value=platform_config,
     ):
-        with patch("nmp.core.models.controllers.backends.k8s_nim_operator.backend.compile_nimservice") as mock_compile:
+        with patch(
+            "nmp.core.models.controllers.backends.k8s_nim_operator.reconcilers.nim_operator.compile_nimservice"
+        ) as mock_compile:
             mock_nimservice = MagicMock()
             mock_nimservice.model_dump.return_value = {"apiVersion": "apps.nvidia.com/v1alpha1", "kind": "NIMService"}
             mock_compile.return_value = mock_nimservice
 
+            _sync_reconcilers(k8s_backend)
             status_update = await k8s_backend.create_model_deployment(
                 sample_deployment, sample_config, files_service_model_entity
             )
@@ -831,11 +932,14 @@ async def test_create_model_deployment_without_sft_model(k8s_backend, sample_dep
     k8s_backend._dynamic_client.resources.get.return_value = mock_nimservice_resource
 
     # Mock the compile_nimservice function
-    with patch("nmp.core.models.controllers.backends.k8s_nim_operator.backend.compile_nimservice") as mock_compile:
+    with patch(
+        "nmp.core.models.controllers.backends.k8s_nim_operator.reconcilers.nim_operator.compile_nimservice"
+    ) as mock_compile:
         mock_nimservice = MagicMock()
         mock_nimservice.model_dump.return_value = {"apiVersion": "apps.nvidia.com/v1alpha1", "kind": "NIMService"}
         mock_compile.return_value = mock_nimservice
 
+        _sync_reconcilers(k8s_backend)
         status_update = await k8s_backend.create_model_deployment(
             sample_deployment, sample_config, non_sft_model_entity
         )
@@ -915,7 +1019,9 @@ async def test_create_model_deployment_with_sft_model_and_revision(k8s_backend, 
         k8s_backend._dynamic_client.resources.get.side_effect = get_resource
 
         # Mock the compile_nimservice function
-        with patch("nmp.core.models.controllers.backends.k8s_nim_operator.backend.compile_nimservice") as mock_compile:
+        with patch(
+            "nmp.core.models.controllers.backends.k8s_nim_operator.reconcilers.nim_operator.compile_nimservice"
+        ) as mock_compile:
             mock_nimservice = MagicMock()
             mock_nimservice.model_dump.return_value = {
                 "apiVersion": "apps.nvidia.com/v1alpha1",
@@ -923,6 +1029,7 @@ async def test_create_model_deployment_with_sft_model_and_revision(k8s_backend, 
             }
             mock_compile.return_value = mock_nimservice
 
+            _sync_reconcilers(k8s_backend)
             status_update = await k8s_backend.create_model_deployment(
                 sample_deployment, sample_config, sft_model_entity
             )
@@ -1016,11 +1123,14 @@ async def test_create_model_deployment_nimcache_uses_fileset_not_entity_name(
         "nmp.core.models.controllers.backends.k8s_nim_operator.nimservice_compiler.get_platform_config",
         return_value=platform_config,
     ):
-        with patch("nmp.core.models.controllers.backends.k8s_nim_operator.backend.compile_nimservice") as mock_compile:
+        with patch(
+            "nmp.core.models.controllers.backends.k8s_nim_operator.reconcilers.nim_operator.compile_nimservice"
+        ) as mock_compile:
             mock_nimservice = MagicMock()
             mock_nimservice.model_dump.return_value = {"apiVersion": "apps.nvidia.com/v1alpha1", "kind": "NIMService"}
             mock_compile.return_value = mock_nimservice
 
+            _sync_reconcilers(k8s_backend)
             status_update = await k8s_backend.create_model_deployment(
                 sample_deployment, sample_config, mismatched_model_entity
             )
@@ -1089,22 +1199,22 @@ class TestFormatDuration:
 
 
 class TestWithRestartInfo:
-    """Tests for K8sNimOperatorServiceBackend._with_restart_info."""
+    """Tests for BaseReconciler._with_restart_info."""
 
     def test_no_restarts(self):
-        assert K8sNimOperatorServiceBackend._with_restart_info("some status", 0) == "some status"
+        assert BaseReconciler._with_restart_info("some status", 0) == "some status"
 
     def test_with_restarts(self):
-        assert K8sNimOperatorServiceBackend._with_restart_info("some status", 3) == "some status, restarts: 3"
+        assert BaseReconciler._with_restart_info("some status", 3) == "some status, restarts: 3"
 
 
 class TestGetPodRestartCount:
-    """Tests for K8sNimOperatorServiceBackend._get_pod_restart_count."""
+    """Tests for BaseReconciler._get_pod_restart_count."""
 
     def test_no_container_statuses(self):
         pod = MagicMock()
         pod.status.container_statuses = None
-        assert K8sNimOperatorServiceBackend._get_pod_restart_count(pod) == 0
+        assert BaseReconciler._get_pod_restart_count(pod) == 0
 
     def test_multiple_containers_returns_max(self):
         pod = MagicMock()
@@ -1113,7 +1223,7 @@ class TestGetPodRestartCount:
         cs2 = MagicMock()
         cs2.restart_count = 7
         pod.status.container_statuses = [cs1, cs2]
-        assert K8sNimOperatorServiceBackend._get_pod_restart_count(pod) == 7
+        assert BaseReconciler._get_pod_restart_count(pod) == 7
 
 
 class TestDeploymentElapsedSeconds:
@@ -1161,10 +1271,11 @@ class TestPendingTimeoutStatusTransition:
         backend._dynamic_client.resources.get.return_value = _make_nimservice_mock("NotReady")
 
         with patch.object(
-            backend,
+            backend._nim_reconciler,
             "_get_pod_status_from_deployment",
             return_value=DeploymentStatusUpdate(status="PENDING", status_message="Waiting", host_url=None),
         ):
+            _sync_reconcilers(backend)
             result = await backend.get_model_deployment_status(sample_deployment)
             assert result.status == "PENDING"
             # No elapsed/timeout in message (stable message to avoid new history entry every poll)
@@ -1179,11 +1290,12 @@ class TestPendingTimeoutStatusTransition:
         backend._dynamic_client.resources.get.return_value = _make_nimservice_mock("NotReady")
 
         with patch.object(
-            backend,
+            backend._nim_reconciler,
             "_get_pod_status_from_deployment",
             return_value=DeploymentStatusUpdate(status="PENDING", status_message="Waiting", host_url=None),
         ):
             with _mock_pod_backend(backend, pod_logs="timeout error log"):
+                _sync_reconcilers(backend)
                 result = await backend.get_model_deployment_status(sample_deployment)
                 assert result.status == "ERROR"
                 assert "timed out" in result.status_message
@@ -1199,6 +1311,7 @@ class TestPendingTimeoutStatusTransition:
         backend._dynamic_client = MagicMock()
         backend._dynamic_client.resources.get.return_value = _make_nimservice_mock("Ready")
 
+        _sync_reconcilers(backend)
         result = await backend.get_model_deployment_status(sample_deployment)
         assert result.status == "READY"
 
@@ -1214,6 +1327,7 @@ class TestPendingTimeoutStatusTransition:
         backend._dynamic_client = MagicMock()
         backend._dynamic_client.resources.get.return_value = _make_nimservice_mock(nim_state)
 
+        _sync_reconcilers(backend)
         result = await backend.get_model_deployment_status(sample_deployment)
         assert result.status != "PENDING"
 
@@ -1228,11 +1342,12 @@ class TestPendingTimeoutStatusTransition:
         backend._dynamic_client.resources.get.return_value = _make_nimservice_mock("NotReady")
 
         with patch.object(
-            backend,
+            backend._nim_reconciler,
             "_get_pod_status_from_deployment",
             return_value=DeploymentStatusUpdate(status="PENDING", status_message="Waiting", host_url=None),
         ):
             with _mock_pod_backend(backend):
+                _sync_reconcilers(backend)
                 result = await backend.get_model_deployment_status(sample_deployment)
                 assert result.status == "ERROR"
                 assert result.error_details["reason"] == "pending_timeout"
@@ -1248,11 +1363,10 @@ class TestCrashLoopDetection:
 
     @pytest.fixture
     def backend(self, mock_nmp_sdk, mock_k8s_config):
-        b = K8sNimOperatorServiceBackend(nmp_sdk=mock_nmp_sdk, config={}, huggingface_model_puller="img:tag")
-        b._backend_config = K8sNimOperatorConfig(pending_timeout_seconds=7200, max_restart_count=5)
-        b._k8s_namespace = "test-ns"
-        b._k8s_client = MagicMock()
-        return b
+        return _status_helper_reconciler(
+            namespace="test-ns",
+            backend_config=K8sNimOperatorConfig(pending_timeout_seconds=7200, max_restart_count=5),
+        )
 
     def test_no_container_statuses(self, backend):
         pod = MagicMock()
@@ -1313,11 +1427,10 @@ class TestPendingTimeoutErrorMessage:
 
     @pytest.fixture
     def backend(self, mock_nmp_sdk, mock_k8s_config):
-        b = K8sNimOperatorServiceBackend(nmp_sdk=mock_nmp_sdk, config={}, huggingface_model_puller="img:tag")
-        b._backend_config = K8sNimOperatorConfig(pending_timeout_seconds=120, max_restart_count=5)
-        b._k8s_namespace = "my-ns"
-        b._k8s_client = MagicMock()
-        return b
+        return _status_helper_reconciler(
+            namespace="my-ns",
+            backend_config=K8sNimOperatorConfig(pending_timeout_seconds=120, max_restart_count=5),
+        )
 
     def test_error_message_with_pod_name(self, backend):
         with _mock_pod_backend(backend, pod_logs="error log tail"):
@@ -1375,11 +1488,12 @@ class TestControllerRestartResilience:
         backend._dynamic_client.resources.get.return_value = _make_nimservice_mock("NotReady")
 
         with patch.object(
-            backend,
+            backend._nim_reconciler,
             "_get_pod_status_from_deployment",
             return_value=DeploymentStatusUpdate(status="PENDING", status_message="Waiting", host_url=None),
         ):
             with _mock_pod_backend(backend):
+                _sync_reconcilers(backend)
                 result = await backend.get_model_deployment_status(sample_deployment)
                 assert result.status == "ERROR"
                 assert result.error_details["reason"] == "pending_timeout"
@@ -1396,10 +1510,11 @@ class TestControllerRestartResilience:
         backend._dynamic_client.resources.get.return_value = _make_nimservice_mock("NotReady")
 
         with patch.object(
-            backend,
+            backend._nim_reconciler,
             "_get_pod_status_from_deployment",
             return_value=DeploymentStatusUpdate(status="PENDING", status_message="Waiting", host_url=None),
         ):
+            _sync_reconcilers(backend)
             result = await backend.get_model_deployment_status(sample_deployment)
             assert result.status == "PENDING"
             # No elapsed/timeout in message (stable message to avoid new history entry every poll)
@@ -1415,10 +1530,7 @@ class TestFetchPodLogs:
 
     @pytest.fixture
     def backend(self, mock_nmp_sdk, mock_k8s_config):
-        b = K8sNimOperatorServiceBackend(nmp_sdk=mock_nmp_sdk, config={}, huggingface_model_puller="img:tag")
-        b._k8s_namespace = "default"
-        b._k8s_client = MagicMock()
-        return b
+        return _status_helper_reconciler(namespace="default")
 
     def test_returns_logs_normally(self, backend):
         with _mock_pod_backend(backend, pod_logs="log line 1\nlog line 2"):
@@ -1434,7 +1546,7 @@ class TestFetchPodLogs:
     def test_returns_empty_on_exception(self, backend):
         mock_core_v1 = MagicMock()
         mock_core_v1.read_namespaced_pod_log.side_effect = Exception("API error")
-        with patch(f"{_K8S_BACKEND_MODULE}.k8s_client.CoreV1Api", return_value=mock_core_v1):
+        with patch(f"{_RECON_BASE_MODULE}.k8s_client.CoreV1Api", return_value=mock_core_v1):
             result = backend._fetch_pod_logs("pod-1")
             assert result == ""
 
@@ -1463,10 +1575,11 @@ class TestAugmentedPendingMessages:
         backend._dynamic_client.resources.get.return_value = _make_nimservice_mock("NotReady")
 
         with patch.object(
-            backend,
+            backend._nim_reconciler,
             "_get_pod_status_from_deployment",
             return_value=DeploymentStatusUpdate(status="PENDING", status_message="Waiting", host_url=None),
         ):
+            _sync_reconcilers(backend)
             result = await backend.get_model_deployment_status(sample_deployment)
             assert result.status == "PENDING"
             # No elapsed/timeout appended (stable message to avoid new history entry every poll)
@@ -1481,6 +1594,7 @@ class TestAugmentedPendingMessages:
         pod = _make_pod(restart_count=3, phase="Running")
 
         with _mock_pod_backend(backend, pod=pod):
+            _sync_reconcilers(backend)
             result = await backend.get_model_deployment_status(sample_deployment)
             assert result.status == "PENDING"
             assert "restarts: 3" in result.status_message
@@ -1494,6 +1608,7 @@ class TestAugmentedPendingMessages:
         pod = _make_pod(restart_count=0, phase="Running")
 
         with _mock_pod_backend(backend, pod=pod):
+            _sync_reconcilers(backend)
             result = await backend.get_model_deployment_status(sample_deployment)
             assert result.status == "PENDING"
             assert "restarts" not in result.status_message
@@ -1509,10 +1624,7 @@ class TestFindPodName:
 
     @pytest.fixture
     def backend(self, mock_nmp_sdk, mock_k8s_config):
-        b = K8sNimOperatorServiceBackend(nmp_sdk=mock_nmp_sdk, config={}, huggingface_model_puller="img:tag")
-        b._k8s_namespace = "default"
-        b._k8s_client = MagicMock()
-        return b
+        return _status_helper_reconciler(namespace="default")
 
     def test_returns_pod_name(self, backend):
         pod = _make_pod(name="found-pod-xyz")
@@ -1531,8 +1643,8 @@ class TestFindPodName:
         mock_core_v1.list_namespaced_pod.return_value = pods_list
 
         with (
-            patch(f"{_K8S_BACKEND_MODULE}.k8s_client.AppsV1Api", return_value=mock_apps_v1),
-            patch(f"{_K8S_BACKEND_MODULE}.k8s_client.CoreV1Api", return_value=mock_core_v1),
+            patch(f"{_RECON_BASE_MODULE}.k8s_client.AppsV1Api", return_value=mock_apps_v1),
+            patch(f"{_RECON_BASE_MODULE}.k8s_client.CoreV1Api", return_value=mock_core_v1),
         ):
             result = backend._find_pod_name("resource-1")
             assert result is None
@@ -1541,7 +1653,7 @@ class TestFindPodName:
         mock_apps_v1 = MagicMock()
         mock_apps_v1.read_namespaced_deployment.side_effect = k8s_client.exceptions.ApiException(status=404)
 
-        with patch(f"{_K8S_BACKEND_MODULE}.k8s_client.AppsV1Api", return_value=mock_apps_v1):
+        with patch(f"{_RECON_BASE_MODULE}.k8s_client.AppsV1Api", return_value=mock_apps_v1):
             result = backend._find_pod_name("resource-1")
             assert result is None
 
@@ -1568,6 +1680,7 @@ class TestCrashLoopPrecedence:
         pod = _make_pod(restart_count=3, waiting_reason="CrashLoopBackOff")
 
         with _mock_pod_backend(backend, pod=pod, pod_logs="crash"):
+            _sync_reconcilers(backend)
             result = await backend.get_model_deployment_status(sample_deployment)
             assert result.status == "ERROR"
             assert result.error_details["reason"] == "crash_loop"
@@ -1593,6 +1706,7 @@ class TestNIMServiceFailedState:
         mock_resource = _make_nimservice_mock("Failed", [{"type": "Failed", "message": "out of GPU memory"}])
         backend._dynamic_client.resources.get.return_value = mock_resource
 
+        _sync_reconcilers(backend)
         result = await backend.get_model_deployment_status(sample_deployment)
         assert result.status == "ERROR"
         assert "NIMService failed" in result.status_message
@@ -1619,6 +1733,7 @@ class TestLostState:
         mock_resource.get.side_effect = k8s_dynamic_exceptions.NotFoundError(MagicMock())
         backend._dynamic_client.resources.get.return_value = mock_resource
 
+        _sync_reconcilers(backend)
         result = await backend.get_model_deployment_status(sample_deployment)
         assert result.status == "LOST"
 
@@ -1652,6 +1767,7 @@ async def test_list_managed_deployment_names_returns_workspace_name_from_labels(
     mock_nimservice_api.get.return_value = list_result
     backend._dynamic_client.resources.get.return_value = mock_nimservice_api
 
+    _sync_reconcilers(backend)
     names = await backend.list_managed_deployment_names()
 
     assert names == ["ws-a/dep1", "ws-b/dep2"]
@@ -1670,6 +1786,7 @@ async def test_list_managed_deployment_names_empty_when_no_resources(mock_nmp_sd
     mock_nimservice_api.get.return_value = list_result
     backend._dynamic_client.resources.get.return_value = mock_nimservice_api
 
+    _sync_reconcilers(backend)
     names = await backend.list_managed_deployment_names()
 
     assert names == []
@@ -1681,7 +1798,7 @@ async def test_delete_model_deployment_by_id_calls_delete_resources(mock_nmp_sdk
     with patch(f"{_K8S_BACKEND_MODULE}.K8sNimOperatorServiceBackend._get_current_namespace", return_value="default"):
         backend = K8sNimOperatorServiceBackend(mock_nmp_sdk, {}, "pull-image")
     backend._k8s_namespace = "default"
-    with patch.object(backend, "_delete_resources_by_model_deployment_id") as mock_delete:
+    with patch.object(backend, "_delete_resources_by_model_deployment_id", new_callable=AsyncMock) as mock_delete:
         mock_delete.return_value = DeploymentStatusUpdate(status="DELETED", status_message="")
         result = await backend.delete_model_deployment("my-ws", "my-name")
     mock_delete.assert_called_once_with("my-ws", "my-name")
@@ -1742,6 +1859,7 @@ async def test_vllm_create_emits_pvc_and_job_only(k8s_backend, sample_deployment
     config = _vllm_config(gpu=2)
 
     with patch.object(backend, "_resolve_model_source", return_value=("default", "qwen", None)):
+        _sync_reconcilers(backend)
         result = await backend.create_model_deployment(sample_deployment, config, None)
 
     assert result.status == "PENDING"
@@ -1762,6 +1880,7 @@ async def test_generic_engine_rejected_on_k8s(k8s_backend, sample_deployment):
     backend = _vllm_backend(k8s_backend)
     config = _vllm_config()
     config.engine = "generic"
+    _sync_reconcilers(backend)
     result = await backend.create_model_deployment(sample_deployment, config, None)
     assert result.status == "ERROR"
     assert "generic" in result.status_message.lower()
@@ -1779,6 +1898,7 @@ async def test_vllm_status_job_running_is_pending(k8s_backend, sample_deployment
     # No serving Deployment yet (still in pull phase).
     backend._apps_v1.read_namespaced_deployment.side_effect = _api_exception(404)
 
+    _sync_reconcilers(backend)
     result = await backend.get_model_deployment_status(sample_deployment, config, None)
     assert result.status == "PENDING"
     assert "weights" in result.status_message.lower()
@@ -1798,6 +1918,7 @@ async def test_vllm_status_job_failed_is_error(k8s_backend, sample_deployment):
     # No serving Deployment yet (failed during pull phase).
     backend._apps_v1.read_namespaced_deployment.side_effect = _api_exception(404)
 
+    _sync_reconcilers(backend)
     result = await backend.get_model_deployment_status(sample_deployment, config, None)
     assert result.status == "ERROR"
     assert result.error_details["reason"] == "weight_pull_failed"
@@ -1822,6 +1943,7 @@ async def test_vllm_status_job_complete_creates_deployment(k8s_backend, sample_d
     created_dep.metadata.uid = "dep-uid"
     backend._apps_v1.create_namespaced_deployment.return_value = created_dep
 
+    _sync_reconcilers(backend)
     result = await backend.get_model_deployment_status(sample_deployment, config, None)
 
     assert result.status == "PENDING"
@@ -1848,6 +1970,7 @@ async def test_vllm_status_p3_waits_for_puller_pod_to_release_volume(k8s_backend
     # Puller pod still terminating -> volume not yet released.
     backend._core_v1.list_namespaced_pod.return_value = MagicMock(items=[MagicMock()])
 
+    _sync_reconcilers(backend)
     result = await backend.get_model_deployment_status(sample_deployment, config, None)
 
     assert result.status == "PENDING"
@@ -1878,7 +2001,8 @@ async def test_vllm_status_job_complete_with_lora_wires_sidecar(k8s_backend, sam
     platform_cfg.image_registry = "my-registry"
     platform_cfg.image_tag = "local"
     platform_cfg.to_shared_envvars.return_value = {"NMP_SHARED": "1"}
-    with patch(f"{_K8S_BACKEND_MODULE}.get_platform_config", return_value=platform_cfg):
+    with patch(f"{_RECON_K8S_MODULE}.get_platform_config", return_value=platform_cfg):
+        _sync_reconcilers(backend)
         result = await backend.get_model_deployment_status(sample_deployment, config, None)
 
     assert result.status == "PENDING"
@@ -1908,6 +2032,7 @@ async def test_vllm_status_job_absent_pvc_present_resumes_p3_not_lost(k8s_backen
     created_dep.metadata.uid = "dep-uid"
     backend._apps_v1.create_namespaced_deployment.return_value = created_dep
 
+    _sync_reconcilers(backend)
     result = await backend.get_model_deployment_status(sample_deployment, config, None)
 
     assert result.status == "PENDING"
@@ -1925,6 +2050,7 @@ async def test_vllm_status_job_and_pvc_absent_is_lost(k8s_backend, sample_deploy
     backend._batch_v1.read_namespaced_job.side_effect = _api_exception(404)
     backend._core_v1.read_namespaced_persistent_volume_claim.side_effect = _api_exception(404)
 
+    _sync_reconcilers(backend)
     result = await backend.get_model_deployment_status(sample_deployment, config, None)
     assert result.status == "LOST"
 
@@ -1944,6 +2070,7 @@ async def test_vllm_status_deployment_ready_is_ready(k8s_backend, sample_deploym
     dep.status.ready_replicas = 1
     backend._apps_v1.read_namespaced_deployment.return_value = dep
 
+    _sync_reconcilers(backend)
     result = await backend.get_model_deployment_status(sample_deployment, config, None)
     assert result.status == "READY"
     assert result.host_url is not None
@@ -1961,7 +2088,8 @@ async def test_vllm_update_unchanged_source_does_not_repull(k8s_backend, sample_
     backend._batch_v1.read_namespaced_job.return_value = existing_job
 
     with patch.object(backend, "_resolve_model_source", return_value=("default", "qwen", None)):
-        with patch.object(backend, "_delete_vllm_resources") as mock_delete:
+        with patch.object(backend._k8s_reconciler, "_delete_vllm_resources") as mock_delete:
+            _sync_reconcilers(backend)
             result = await backend.update_model_deployment(sample_deployment, config, None)
 
     mock_delete.assert_not_called()
@@ -1981,7 +2109,8 @@ async def test_vllm_update_changed_source_repulls(k8s_backend, sample_deployment
     backend._batch_v1.read_namespaced_job.return_value = existing_job
 
     with patch.object(backend, "_resolve_model_source", return_value=("default", "qwen", "v2")):
-        with patch.object(backend, "_delete_vllm_resources") as mock_delete:
+        with patch.object(backend._k8s_reconciler, "_delete_vllm_resources") as mock_delete:
+            _sync_reconcilers(backend)
             result = await backend.update_model_deployment(sample_deployment, config, None)
 
     mock_delete.assert_called_once()
@@ -2005,5 +2134,6 @@ async def test_vllm_list_managed_unions_deployments(k8s_backend):
     }
     backend._apps_v1.list_namespaced_deployment.return_value = MagicMock(items=[dep])
 
+    _sync_reconcilers(backend)
     names = await backend.list_managed_deployment_names()
     assert "default/qwen" in names
